+++ tank_module/include/UltrasonicLevelSensor.hpp
@@ -6,8 +6,8 @@
 class UltrasonicLevelSensor
 {
 private:
-    static constexpr int trigPin = 4;               // Pino do TRIG
-    static constexpr int echoPin = 2;               // Pino do ECHO
+    static constexpr u_int8_t trigPin = 4;               // Pino do TRIG
+    static constexpr u_int8_t echoPin = 2;               // Pino do ECHO
     static constexpr float tankHeight = 50.0f;      // Altura máxima do reservatório (em cm)
     static constexpr float tankOffSet = 10.0f;      // Offset para ajuste na medição
     static constexpr float soundSpeed = 0.0343f;    // Velocidade do som no ar (em cm/us)
